`buy` command fixes / additional flexibility
When specifying an item that didn't exactly match one of the keys in weightDict, was throwing a divide by zero error. 
Fix: bundle = matchItem.get("bundle", 1) if typeof(matchItem) == 'SafeDict' else 1

Improvement: Added function to grab a list of fuzzy/partial matches. The `buy` command will now only do the buy if it has a match, and show some options if an exact match couldn't be found. "buy dag" will work to buy a dagger, "buy sword" will list the items that contain the word 'sword'.