Fix for `!bag buy`
When attempting to buy something you can't afford, it would throw a `bag is not defined` error, due to the `bag` variable only getting set if the purchase was done successfully. Defaulting it to None fixes the problem.

Giving more information in the output of the buy command, including the total cost.